+++ src/utils/parsing.h
@@ -35,7 +35,7 @@ void write_ip4(std::ostream& s, uint32_t ip);
 
 uint32_t parse_ip4(const char* s);
 
-void write_ip6(std::ostream& s, uint32_t ip);
+void write_ip6(std::ostream& s, uint64_t ip);
 
 uint32_t parse_ip6(const char* s);
